+++ src/nrniv/nrncore_write/utils/nrncore_utils.h
@@ -1,13 +1,16 @@
-#ifndef NRN_CORENRN_UTILS_H
-#define NRN_CORENRN_UTILS_H
+#ifndef NRN_NRNCORE_UTILS_H
+#define NRN_NRNCORE_UTILS_H
 
 #include <string>
 extern "C" {
+
+class NrnThread;
 
 void model_ready();
 int count_distinct(double *data, int len);
 extern void nrnbbcore_register_mapping();
 bool file_exist(const std::string& path);
+int nrn_dblpntr2nrncore(double* pd, NrnThread& nt, int& type, int& index);
 
 
 #if defined(HAVE_DLFCN_H)
@@ -21,4 +24,5 @@ void check_coreneuron_compatibility(void* handle);
 
 }
 
-#endif //NRN_CORENRN_UTILS_H
+#endif //NRN_NRNCORE_UTILS_H
+
